services_LLM/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de datos PostgreSQL
psycopg[binary]==3.2.3
psycopg[pool]==3.2.3

# LangChain - Core y componentes modernos
langchain==0.3.7
langchain-core==0.3.15
langchain-postgres==0.0.12
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2

# Vector store
pgvector>=0.2.5,<0.3.0

# Procesamiento de PDF
pypdf==5.1.0

# Embeddings
sentence-transformers==3.3.1
huggingface-hub==0.26.2

# Cliente HTTP para DeepSeek API
openai==1.55.3
httpx==0.27.2

# Utilidades
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Logging y monitoreo (opcional)
loguru==0.7.3
